Convert silt to TypeScript
Remove some modules that were no longer required and difficult to type
correctly. They were primarily used to embed observables in JSX, but we
removed that functionality for a later release.
