+++ babel.config.js
@@ -1,15 +1,18 @@
 module.exports = api => {
-    api.cache(false);
+  api.cache(false);
 
-    return {
-        plugins: [
-            '@babel/plugin-proposal-class-properties',
-            'ramda'
-        ].filter(Boolean),
-        presets: [
-            'brookjs',
-            ['@babel/env', {
-                modules: process.env.NODE_ENV === 'test' ? 'commonjs' : false
-            }]]
-    };
+  return {
+    plugins: ['@babel/plugin-proposal-class-properties', 'ramda'].filter(
+      Boolean
+    ),
+    presets: [
+      'brookjs',
+      [
+        '@babel/env',
+        {
+          modules: process.env.NODE_ENV === 'test' ? 'commonjs' : false
+        }
+      ]
+    ]
+  };
 };
