+++ babel.config.ts.js
@@ -1,17 +1,12 @@
 module.exports = api => {
-    api.cache(false);
+  api.cache(false);
 
-    return {
-        plugins: [
-            '@babel/plugin-proposal-class-properties',
-            'ramda',
-            process.env.NODE_ENV === 'test' && '@babel/transform-modules-commonjs'
-        ].filter(Boolean),
-        presets: ['brookjs', '@babel/typescript', '@babel/env'],
-        env: {
-            test: {
-                plugins: ['@babel/transform-modules-commonjs']
-            }
-        }
-    };
+  return {
+    plugins: [
+      '@babel/plugin-proposal-class-properties',
+      'ramda',
+      process.env.NODE_ENV === 'test' && '@babel/transform-modules-commonjs'
+    ].filter(Boolean),
+    presets: ['brookjs', '@babel/typescript', '@babel/env', '@babel/react']
+  };
 };
