+++ package.json
@@ -14,6 +14,8 @@
     "@types/kefir": "^3.8.0",
     "@types/mocha": "^5.2.5",
     "@types/node": "^10.12.18",
+    "@types/react": "^16.8.0",
+    "@types/recompose": "^0.30.5",
     "@types/redux": "^3.6.0",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
@@ -29,8 +31,6 @@
     "cucumber": "^5.0.0",
     "dirty-chai": "^2.0.1",
     "docz": "^0.11.0",
-    "enzyme": "^3.4.4",
-    "enzyme-adapter-react-16.3": "^1.1.0",
     "eslint": "^5.0.0",
     "eslint-config-valtech": "^1.0.0",
     "eslint-import-resolver-node": "^0.3.1",
@@ -45,7 +45,8 @@
     "lerna": "^3.0.6",
     "prettier": "^1.15.3",
     "raf": "^3.4.0",
-    "react-dom": "^16.2.0",
+    "react-dom": "^16.8.0",
+    "react-testing-library": "^6.0.0",
     "redux-mock-store": "^1.5.3",
     "rimraf": "^2.5.4",
     "rollup": "^0.66.0",
