+++ packages/brookjs-silt/.babelrc
@@ -1,3 +1,3 @@
 {
-  "extends": "../../babel.config"
+  "extends": "../../babel.config.ts"
 }
