+++ packages/brookjs-silt/.beaverrc.js
@@ -4,26 +4,27 @@ export const dir = 'src';
  * Mocha testing configuration.
  */
 export const mocha = {
-    reporter: 'spec',
-    ui: 'bdd',
-    requires: [
-        '@babel/register',
-        'react-testing-library',
-        'jsdom-global/register',
-        'raf/polyfill',
-    ]
+  reporter: 'spec',
+  ui: 'bdd',
+  requires: [
+    'ts-node/register/transpile-only',
+    '@babel/register',
+    'react-testing-library',
+    'jsdom-global/register',
+    'raf/polyfill'
+  ]
 };
 
 /**
  * Webpack build configuration.
  */
 export const webpack = {
-    entry: {
-        app: 'app.js'
-    },
-    output: {
-        path: 'dist/',
-        filename: '[name].js'
-    },
-    modifier: x => x
+  entry: {
+    app: 'app.js'
+  },
+  output: {
+    path: 'dist/',
+    filename: '[name].js'
+  },
+  modifier: x => x
 };
