--- packages/brookjs-silt/src/context.js
@@ -1,6 +1,0 @@
-import createReactContext from 'create-react-context';
-import Kefir from 'kefir';
-
-const { Provider, Consumer } = createReactContext(Kefir.pool());
-
-export { Provider, Consumer };
