+++ packages/brookjs-silt/src/index.ts
@@ -1,8 +1,6 @@
-import loop from './loop';
-import observableOf from './observableOf';
 import RootJunction from './RootJunction';
 import toJunction from './toJunction';
 import view from './view';
 import withRef$ from './withRef';
 
-export { loop, observableOf, RootJunction, toJunction, view, withRef$ };
+export { RootJunction, toJunction, view, withRef$ };
