+++ rollup.config.js
@@ -12,7 +12,6 @@ const bbl = {
 
 const cjs = {
   namedExports: {
-    'create-react-ref': ['createRef', 'forwardRef', 'getRef'],
     '@storybook/addon-actions': ['action', 'ADDON_ID', 'PANEL_ID', 'EVENT_ID'],
     '@storybook/addons': ['makeDecorator'],
     react: [
