+++ rollup.config.ts
@@ -19,7 +19,6 @@ const ts = {
 
 const cjs = {
   namedExports: {
-    'create-react-ref': ['createRef', 'forwardRef', 'getRef'],
     '@storybook/addon-actions': ['action', 'ADDON_ID', 'PANEL_ID', 'EVENT_ID'],
     '@storybook/addons': ['makeDecorator'],
     react: [
