+++ tsconfig.build.json
@@ -4,7 +4,7 @@
     "allowJs": false,
     "skipLibCheck": false,
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": false,
+    "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "module": "esnext",
