+++ tsconfig.json
@@ -4,7 +4,7 @@
     "allowJs": false,
     "skipLibCheck": false,
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": false,
+    "allowSyntheticDefaultImports": true,
     "strict": true,
     "forceConsistentCasingInFileNames": true,
     "module": "commonjs",
@@ -13,7 +13,7 @@
     "isolatedModules": true,
     "noEmit": true,
     "declaration": false,
-    "jsx": "preserve"
+    "jsx": "react"
   },
   "include": ["packages/**/*.ts"]
 }
